pyproject.toml
[project]
name = "treeland-autogui-mcp"
version = "0.1.0"
description = "Automatic operation of on-screen GUI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "Pillow>=10.4.0",
    "pyautogui",
    "uvicorn>=0.34.0",
    "pyperclip>=1.9.0",
    "mcp[cli]>=1.3.0",
    "requests>=2.32.3",
    "langchain-google-genai>=4.2.1",
    "google-generativeai>=0.8.6",
    "langchain-community>=0.4.1",
    "langchain>=1.2.15",
    "langchain-core>=1.2.28",
    "langchain-openai>=1.1.12",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-mcp>=0.2.1",
    "google-genai>=1.71.0",
]
license = {text = "MIT License"}
license-files = ["LICENSE"]

[project.scripts]
treeland-autogui-mcp = "mcp_autogui:main"

[tool.hatch.build.targets.wheel]
packages = [
  "src/mcp_autogui"
]

[project.optional-dependencies]
langchain = [
    "langchain-mcp",
    "langchain",
    "langchain-community",
    "langchain-google-genai",
    "langchain-openai>=0.3.6",
    "langgraph",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
langchain-mcp = { git = "https://github.com/NON906/langchain-mcp.git", branch = "main" }
pyautogui = { git = "https://github.com/zorowk/pyautogui.git", branch = "master" }
